refactor: move client into own file
